requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
pymupdf>=1.23.0
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
